requirements/base.txt
fastapi[standard]
uvicorn
motor
pydantic
pydantic-settings
python-dotenv
passlib[bcrypt]
python-multipart
python-jose[cryptography]
fastapi-mail
pyopenssl
cryptography
bcrypt==4.0.1
passlib[bcrypt]==1.7.4
celery
redis
confluent-kafka
sentence-transformers
google-generativeai
